cdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ndl-ocr-lite-cdk"
version = "1.0.0"
description = "CDK infrastructure for NDL-OCR Lite Lambda"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.240.0",
    "aws-cdk.aws-bedrock-agentcore-alpha>=2.240.0a0",
    "constructs>=10.0.0",
]

[tool.setuptools]
packages = ["stacks"]
